# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Character AI API
  description: "\n# Character AI API\n\nA comprehensive API for managing interactive\
    \ AI characters in toys and devices.\n\n## Features\n\n- **Character Management**:\
    \ Create, configure, and manage AI characters\n- **Real-time Interaction**: Process\
    \ audio input and generate character responses\n- **Session Memory**: Maintain\
    \ conversation context with configurable limits\n- **Safety & Security**: Content\
    \ filtering and safety analysis\n- **Voice Management**: Handle voice artifacts\
    \ and embeddings\n- **Authentication**: Device identity and JWT-based authentication\n\
    \n## Authentication\n\nThe API supports multiple authentication methods:\n\n1.\
    \ **Device Registration**: Register a new device and receive authentication tokens\n\
    2. **JWT Tokens**: Use Bearer tokens for authenticated requests\n3. **Admin Tokens**:\
    \ Special tokens for administrative operations\n\n## Rate Limiting\n\n- Default:\
    \ 60 requests per minute per device\n- Burst allowance: 10 requests\n- Rate limit\
    \ headers are included in responses\n\n## Safety & Privacy\n\n- All content is\
    \ analyzed for toxicity and PII\n- Child-safe content filtering\n- Configurable\
    \ safety thresholds\n- Privacy-first design with on-device processing\n\n## Character\
    \ Profiles\n\nCharacters are defined using YAML profiles with:\n- Personality\
...
    \        "
  version: 1.0.0
servers:
...
paths:
...
  /performance/compliance:
    get:
      tags:
      - performance
      summary: Check Budget Compliance
      description: Check budget compliance for components.
      operationId: check_budget_compliance_performance_compliance_get
      parameters:
      - name: component
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Component
      - name: time_window_seconds
        in: query
        required: false
        schema:
          type: number
          default: 3600
          title: Time Window Seconds
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Check Budget Compliance Performance Compliance Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /performance/test:
    post:
      tags:
      - performance
      summary: Run Performance Test
      description: Run a performance test for a component.
      operationId: run_performance_test_performance_test_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PerformanceTestRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Run Performance Test Performance Test Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /performance/load-test:
    post:
      tags:
      - performance
      summary: Run Load Test
      description: Run a load test with increasing load.
      operationId: run_load_test_performance_load_test_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoadTestRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Run Load Test Performance Load Test Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /performance/reports:
    get:
      tags:
      - performance
      summary: Get Performance Reports
      description: Get comprehensive performance reports.
      operationId: get_performance_reports_performance_reports_get
      parameters:
      - name: time_window_seconds
        in: query
        required: false
        schema:
          type: number
          default: 3600
          title: Time Window Seconds
      - name: format
        in: query
        required: false
        schema:
          type: string
          default: json
          title: Format
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Performance Reports Performance Reports Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoadTestRequest:
      properties:
        component:
          type: string
          title: Component
        operation:
          type: string
          title: Operation
        initial_load:
          type: integer
          title: Initial Load
          default: 1
        max_load:
          type: integer
          title: Max Load
          default: 10
        step:
          type: integer
          title: Step
          default: 1
        duration_per_step:
          type: number
          title: Duration Per Step
          default: 5.0
        metadata:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Metadata
      type: object
      required:
      - component
      - operation
      title: LoadTestRequest
      description: Request model for load testing.
...
    PerformanceTestRequest:
      properties:
        component:
          type: string
          title: Component
        operation:
          type: string
          title: Operation
        concurrency:
          type: integer
          title: Concurrency
          default: 1
        duration_seconds:
          type: number
          title: Duration Seconds
          default: 10.0
        metadata:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Metadata
      type: object
      required:
      - component
      - operation
      title: PerformanceTestRequest
      description: Request model for performance testing.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from device registration
    AdminToken:
      type: apiKey
      in: header
      name: x-admin-token
      description: Admin token for administrative operations
...
security:
- BearerAuth: []
- AdminToken: []
...